Move isolated network tests to their own module
