+++ zebra-network/src/isolated.rs
@@ -17,6 +17,9 @@ use crate::{
     BoxError, Config, Request, Response,
 };
 
+#[cfg(test)]
+mod tests;
+
 /// Use the provided data stream to create a Zcash connection completely
 /// isolated from all other node state.
 ///
@@ -71,64 +74,4 @@ where
         },
     )
     .map_ok(|client| BoxService::new(client.map_err(Into::into)))
-}
-
-#[cfg(test)]
-mod tests {
-
-    use super::*;
-
-    #[tokio::test]
-    async fn connect_isolated_sends_minimally_distinguished_version_message() {
-        use std::net::SocketAddr;
-
-        use futures::stream::StreamExt;
-        use tokio_util::codec::Framed;
-
-        use crate::{
-            protocol::external::{AddrInVersion, Codec, Message},
-            types::PeerServices,
-        };
-
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let listen_addr = listener.local_addr().unwrap();
-
-        let fixed_isolated_addr: SocketAddr = "0.0.0.0:8233".parse().unwrap();
-
-        let conn = tokio::net::TcpStream::connect(listen_addr).await.unwrap();
-
-        tokio::spawn(connect_isolated(conn, "".to_string()));
-
-        let (conn, _) = listener.accept().await.unwrap();
-
-        let mut stream = Framed::new(conn, Codec::builder().finish());
-        if let Message::Version {
-            services,
-            timestamp,
-            address_from,
-            user_agent,
-            start_height,
-            relay,
-            ..
-        } = stream
-            .next()
-            .await
-            .expect("stream item")
-            .expect("item is Ok(msg)")
-        {
-            // Check that the version message sent by connect_isolated
-            // has the fields specified in the Stolon RFC.
-            assert_eq!(services, PeerServices::empty());
-            assert_eq!(timestamp.timestamp() % (5 * 60), 0);
-            assert_eq!(
-                address_from,
-                AddrInVersion::new(fixed_isolated_addr, PeerServices::empty()),
-            );
-            assert_eq!(user_agent, "");
-            assert_eq!(start_height.0, 0);
-            assert!(!relay);
-        } else {
-            panic!("handshake did not send version message");
-        }
-    }
 }
