+++ zebra-network/src/isolated/tests.rs
@@ -1,0 +1,3 @@
+//! Tests for isolated Zebra connections.
+
+mod vectors;
